Asynchronous scene rendering
Load up lasgun in a webworker and send it a render request for the simple scene

Arbitrary scenes coming soon
