+++ js/www/index.html
@@ -6,6 +6,6 @@
   </head>
   <body>
     <canvas id="output" width="512" height="512"></canvas>
-    <script src="./bootstrap.js"></script>
+    <script src="app.js"></script>
   </body>
 </html>
