+++ js/www/webpack.config.js
@@ -1,14 +1,26 @@
-const CopyWebpackPlugin = require("copy-webpack-plugin");
-const path = require('path');
+const CopyWebpackPlugin = require("copy-webpack-plugin")
+const path = require('path')
 
-module.exports = {
-  entry: "./bootstrap.js",
+const browserConfig = {
+  entry: './app.js',
   output: {
     path: path.resolve(__dirname, "dist"),
-    filename: "bootstrap.js",
+    filename: "app.js",
   },
   mode: "development",
   plugins: [
     new CopyWebpackPlugin(['index.html'])
   ],
-};
+}
+
+const workerConfig = {
+  entry: "./worker.js",
+  target: 'webworker',
+  output: {
+    path: path.resolve(__dirname, "dist"),
+    filename: "worker.js"
+  },
+  mode: "development",
+}
+
+module.exports = [browserConfig, workerConfig]
